# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cyclefi API Specification
  version: 1.0.0
  description: Dokumentasi API Cyclefi (Bahasa Indonesia)
...
paths:
...
  /api/users/current:
    get:
      summary: Mendapatkan data user saat ini
      description: |
        Mengambil data profil user yang sedang login, termasuk relasi phones dan addresses.
      tags: [User]
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token JWT
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: Accept-Language
          in: header
          required: false
          description: Pengaturan bahasa (en, id) dengan callback ke bahasa Inggris.
          schema:
            type: string
            example: en
      responses:
        "200":
          description: Get current user successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    { type: string, example: "Get current user successful" }
                  data:
                    type: object
                    properties:
                      id: { type: integer, example: 1 }
                      fullname: { type: string, example: "John Doe" }
                      username: { type: string, example: "johndoe" }
                      email: { type: string, example: "john@example.com" }
                      is_email_verified: { type: boolean, example: true }
                      email_verified_at:
                        {
                          type: string,
                          format: date-time,
                          example: "2025-08-12T10:00:00Z",
                        }
                      profile_picture:
                        {
                          type: string,
                          example: "http://.../assets/profiles/johndoe.png",
                        }
                      is_active: { type: boolean, example: true }
                      phones:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer, example: 1 }
                            number: { type: string, example: "08123456789" }
                      addresses:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer, example: 1 }
                            address_name: { type: string, example: "Rumah" }
                            address:
                              { type: string, example: "Jl. Mawar No. 1" }
                            latitude: { type: number, example: -6.9 }
                            longitude: { type: number, example: 107.6 }
                            city: {type: string, example: "Unaaha"}
                            state: {type: string, example: "Sulawesi Tenggara"}
                            country: {type: string, example: "Indonesia"}
                            country_code: {type: string, example: "ID"}
                            zipcode: {type: string, example: "93418"}
                            created_at: {type: string, example: "2025-08-14T00:50:59.220Z"}
                            updated_at: {type: string, example: "2025-08-14T00:50:59.220Z"}
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null

    patch:
      summary: Update data profil user saat ini
      description: |
        Mengupdate data profil user (fullname, username, password). Hanya field yang dikirim yang diupdate.
        Untuk update password, wajib mengirim oldPassword, password baru, dan confirmPassword.
      tags: [User]
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token JWT
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: Accept-Language
          in: header
          required: false
          description: Pengaturan bahasa (en, id) dengan callback ke bahasa Inggris.
          schema:
            type: string
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname: { type: string, example: "John Doe" }
                username: { type: string, example: "johndoe" }
                oldPassword:
                  { type: string, format: password, example: "oldpassword123" }
                password:
                  { type: string, format: password, example: "newpassword123" }
                confirmPassword:
                  { type: string, format: password, example: "newpassword123" }
              description: Kirim hanya field yang ingin diupdate. Untuk update password, wajib oldPassword, password, confirmPassword.
      responses:
        "200":
          description: Update current user successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    { type: string, example: "Update current user successful" }
                  data:
                    type: object
                    example:
                      fullname: "John Doe"
                      username: "johndoe"
        "400":
          description: Validasi gagal atau old password salah
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: "Validation failed" }
                  errors:
                    {
                      type: object,
                      example: { oldPassword: ["Old password is incorrect"] },
                    }
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null
  /api/users/current/profile-picture:
    patch:
      summary: Upload foto profil user saat ini
      description: |
        Mengganti foto profil user yang sedang login. File gambar dikirim sebagai form-data dengan field profile_picture
      tags:
        - User
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token JWT
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: Accept-Language
          in: header
          required: false
          description: Pengaturan bahasa (en, id) dengan callback ke bahasa Inggris.
          schema:
            type: string
            example: en
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: multipart/form-data
          description: Tipe konten yang dikirim dalam format multipart/form-data karena ada file yang disertakan
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                profile_picture:
                  type: string
                  format: binary
      responses:
        "200":
          description: Foto profil berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Update profile picture successful"
                  data:
                    type: object
                    properties:
                      profile_picture:
                        type: string
                        example: "http://.../assets/profiles/user_1_1755334590343.jpg"
        "400":
          description: Validasi gagal atau file tidak ditemukan
          content:
            application/json:
              example:
                success: false
                message: "No file uploaded"
                errors: null
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null
  /api/users/current/addresses:
    get:
      summary: Mendapatkan daftar alamat user saat ini
      description: |
        Mengambil data alamat user yang sedang login
      tags:
        - Address
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token JWT
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: Accept-Language
          in: header
          required: false
          description: Pengaturan bahasa (en, id) dengan callback ke bahasa Inggris.
          schema:
            type: string
            example: en
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
          description: Tipe konten yang dikirim dalam format application/json
      responses:
        "200":
          description: "Berhasil mendapatkan daftar alamat pengguna saat ini"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Berhasil mendapatkan daftar alamat pengguna saat ini"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        address_name:
                          type: string
                          example: "Rumah Utama"
                        address:
                          type: string
                          example: "Jl. Diponegoro No. 1, Bandung, Jawa Barat, Indonesia"
                        latitude:
                          type: number
                          example: -6.90389
                        longitude:
                          type: number
                          example: 107.61861
                        city:
                          type: string
                          example: "Bandung"
                        state:
                          type: string
                          example: "Jawa Barat"
                        country:
                          type: string
                          example: "Indonesia"
                        country_code:
                          type: string
                          example: "ID"
                        zipcode:
                          type: string
                          example: "40115"
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null
        "404":
          description: Address tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: "Address not found" }

    post:
      summary: Tambah alamat baru untuk user saat ini
      description: |
        Endpoint untuk menambah alamat baru milik user yang sedang login. 
        Sistem akan otomatis melakukan geocoding untuk mendapatkan latitude, longitude, dan detail lokasi dari field `address` yang diberikan.
      tags:
        - Address
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: JWT token (Bearer)
        - name: Accept-Language
          in: header
          required: false
          schema:
            type: string
          description: Bahasa response (en/id)
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
          description: Tipe konten yang dikirim dalam format application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addressName:
                  type: string
                  maxLength: 255
                  example: "Rumah Utama"
                address:
                  type: string
                  maxLength: 255
                  example: "Jl. Diponegoro No. 1, Bandung"
              required:
                - address
      responses:
        "201":
          description: Alamat berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Create address successful."
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      address_name:
                        type: string
                        example: "Rumah Utama"
                      address:
                        type: string
                        example: "Jl. Diponegoro No. 1, Bandung, Jawa Barat, Indonesia"
                      latitude:
                        type: number
                        example: -6.90389
                      longitude:
                        type: number
                        example: 107.61861
                      city:
                        type: string
                        example: "Bandung"
                      state:
                        type: string
                        example: "Jawa Barat"
                      country:
                        type: string
                        example: "Indonesia"
                      country_code:
                        type: string
                        example: "ID"
                      zipcode:
                        type: string
                        example: "40115"
        "400":
          description: Validasi gagal atau geocoding gagal
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Validation failed"
                  errors:
                    type: object
                    example: { "address": ["Field address is required"] }
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null
...
  /api/users/current/phones:
    get:
      summary: Mendapatkan daftar nomor kontak user saat ini
      description: |
        Mengambil seluruh daftar nomor kontak (phone) milik user yang sedang login.
        - Hanya dapat diakses oleh user yang sudah login (dengan Authorization token).
      tags: [Phone]
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token JWT
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: Accept-Language
          in: header
          required: false
          description: Pengaturan bahasa (en, id) dengan callback ke bahasa Inggris.
          schema:
            type: string
            example: en
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
          description: Tipe konten yang dikirim dalam format application/json
      responses:
        "200":
          description: Daftar nomor kontak user berhasil diambil
          content:
            application/json:
              example:
                success: true
                message: Berhasil mendapatkan daftar nomor telepon.
                data:
                  - id: 1
                    number: "08123456789"
                  - id: 2
                    number: "08987654321"
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null
        "404":
          description: Address tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: "Address not found" }

    post:
      summary: Tambah nomor kontak baru untuk user saat ini
      description: |
        Menambahkan nomor kontak (phone) baru untuk user yang sedang login.
        - Nomor telepon tidak boleh duplikat untuk user yang sama.
      tags: [Phone]
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token JWT
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: Accept-Language
          in: header
          required: false
          description: Pengaturan bahasa (en, id) dengan callback ke bahasa Inggris.
          schema:
            type: string
            example: en
      requestBody:
        required: true
        content:
          application/json:
            example:
              number: "08123456789"
      responses:
        "201":
          description: Nomor kontak berhasil ditambahkan
          content:
            application/json:
              example:
                success: true
                message: Berhasil membuat nomor telepon.
                data:
                  id: 3
                  number: "08123456789"
                  user_id: 1
        "400":
          description: Validasi gagal atau nomor telepon sudah ada
          content:
            application/json:
              examples:
                validasi:
                  value:
                    success: false
                    message: Validasi gagal
                    errors:
                      number: "Field number wajib diisi"
                duplikat:
                  value:
                    success: false
                    message: Nomor telepon sudah ditambahkan
                    errors: null
        "401":
          description: Unauthorized (token tidak valid/expired)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized. Need Authorization Token"
                  errors:
                    type: object
                    properties:
                      token:
                        type: string
                        example: null
...
  /api/users/current/notifications/{notificationId}/read:
    patch:
      tags:
        - Notifikasi
      summary: Tandai notifikasi sebagai sudah dibaca
      description: "Tandai notifikasi tertentu milik user saat ini sebagai sudah dibaca (`is_read: true`)"
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: integer
          description: ID notifikasi
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: JWT token (Bearer)
        - name: Accept-Language
          in: header
          required: false
          schema:
            type: string
          description: Bahasa response (en/id)
      responses:
        '200':
          description: Notifikasi berhasil ditandai sebagai sudah dibaca
          content:
            application/json:
              example:
                success: true
                message: "Tandai notifikasi sebagai berhasil dibaca"
                data:
                  count: 1
        '404':
          description: Notifikasi tidak ditemukan
        '401':
          description: Tidak terautentikasi
        '500':
          description: Kesalahan server internal
...